+++ src/components/layout/tenantbar.jsx
@@ -53,10 +53,17 @@ export default function BottomNavbarSidebar({ children }) {
                 to="/resiix/report-issue"
                 sx={{ width: "70px" }}
               />
+
+<ListItemWithLink
+                label="Active Requests"
+                icon={<img src={repair} alt="Custom Icon" />}
+                to="/resiix/reported"
+                sx={{ width: "70px" }}
+              />
               <ListItemWithLink
                 label="History"
                 icon={<DashboardIcon />}
-                to="/resiix/rating"
+                to="/resiix/reportedhistory"
                 sx={{ width: "70px" }}
               />
             </HorizontalList>
